// assets/Scripts/Model/SkillModel.ts
export class SkillCfg {
    ClassName?: string = 'Skill';
    /** id */
    Id: number;
    Name: string;
    Quality: number;
    Res: string;
    Bullet: string;
    Type: number;
    Scale: number;
    Angle: number;
    Speed: number;
    Range: number;
    Cd: number;
    Damage: number;
    DamageRange: number;
    DamageNum: number;
    Damage_interval: number;
    Salvo: number;
    BulletCount: number;
    IntervalAngle: number;
}