// assets/Scripts/Model/SkillTargetModel.ts
export class SkillTargetCfg {
    ClassName?: string = 'SkillTarget';
    Id: number;
    type: Array<number>;
    stress: number;
    random: string;
    targetCamp: number;
    targetHp: Array<number>;
    targetStatus: Array<number>;
    condition: Array<number>;
    notargetStatus: Array<number>;
}