// assets/Scripts/Model/SkillUpgradeMode.ts
export class SkillUpgradeCfg {
    ClassName?: string = 'SkillUpgrade';
    Id: number;
    name: string;
    desc: string;
    icon: string;
    type: number;
    premise: Array<number>;
    aim_skill: Array<number>;
    options_num: number;
    quality: number;
    hp_up: number;
    exp_up: number;
    gold_up: number;
    cd_shorten: number;
    damageRange_up: number;
    damageNum_up: number;
    damage_up: number;
    range_up: number;
    speed_up: number;
    salvo_up: number;
    intensity_up: number;
    unlock: Array<number>;
}